Added conditional rendering for settings button
