--- src/components/Options/Options.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Options = () => {
-  return ( 
-      <div>
-          <p>Options</p>
-      </div>
-   );
-}
- 
-export default Options;
