--- src/components/Options/OptionsButtons.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const OptionsButtons = () => {
-  return ( 
-      <div>
-          <p>OptionsButtons</p>
-      </div>
-   );
-}
- 
-export default OptionsButtons;
