+++ src/components/Verb/Settings/Settings.js
@@ -1,0 +1,38 @@
+import React, { Component } from 'react';
+import VerbTenses from './VerbTenses';
+import NumberPerson from './VerbNumberPerson';
+import './settings.css';
+
+class Settings extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      updateSettings: false
+    };
+  }
+  changeSettings = event => {
+    this.setState({
+      updateSettings: !this.state.updateSettings
+    });
+  };
+
+  render() {
+    const hide = 'settings-buttons-hide';
+    const display = 'settings-buttons-display';
+    return (
+      <div>
+        <div className={this.state.updateSettings ? display : hide}>
+          <VerbTenses updateVerbTenses={this.props.updateVerbTenses} />
+          <NumberPerson updateNumPerson={this.props.updateNumPerson} />
+        </div>
+        <div>
+          <button onClick={this.changeSettings} className="button-options">
+            {this.state.updateSettings ? 'Update Settings' : 'Update Settings?'}
+          </button>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default Settings;
