+++ src/components/Verb/Settings/settings.css
@@ -1,0 +1,7 @@
+.settings-buttons-display {
+  display: inline;
+}
+
+.settings-buttons-hide {
+  display: none;
+}
